Укрупнил компоненты
