+++ src/utils/offer.ts
@@ -1,0 +1,19 @@
+import { Offer } from '../types/offer';
+
+function getFavoriteOffers(offers: Offer[]): Offer[] {
+  return offers.filter(({ isFavorite }) => isFavorite);
+}
+
+function getSortingCityNames(offers: Offer[]): string[] {
+  const cityNames = offers.map(({ city }) => city.name);
+  cityNames.sort((firstCityName, secondCityName) => firstCityName.localeCompare(secondCityName));
+  const sortingCityNames = new Set(cityNames);
+
+  return Array.from(sortingCityNames);
+}
+
+/*
+cityOffers = favoriteOffers.filter(({ city }) => city.name === cityName)
+ */
+
+export { getFavoriteOffers, getSortingCityNames };
